Card Styling
I have the three cards, and have styled the first one, so the others will need to be clones of this with the appropriate colors.
Also, they will need to be able to ALL change on hover, at the minute they all work individually, but work.
There needs to be 7 cards in total !!